Upgrade to the new version of AppSystem

+++ src/Factory/PieceFactory.php
@@ -13,14 +13,14 @@ final class PieceFactory implements PieceFactoryInterface
     /**
      * @throws InvalidStructureException
      */
-    public function create(array $customFields, int $totalWeight): array
+    public function create(array $customFields, float $totalWeight): array
     {
         return (new Piece())
             ->setType(Piece::TYPE_PACKAGE)
             ->setWidth($customFields[Defaults::PACKAGE_WIDTH])
             ->setHeight($customFields[Defaults::PACKAGE_HEIGHT])
             ->setLength($customFields[Defaults::PACKAGE_DEPTH])
-            ->setWeight($totalWeight)
+            ->setWeight((int) $totalWeight)
             ->setQuantity(1)
             ->setNonStandard(false)
             ->structure();
